Update contextual-tabs.md
I've upgraded Fluent.Ribbon from 3.6.1.236 to 6.0.0.208. I had code working where the contextual tab was using "Hidden" and after the upgrade this didn't work anymore. The contextual tab was always visible. This is because only Collapsed is now supported and not anymore Hidden. Might be useful documentation for some other users.